+++ const-str/src/str.rs
@@ -1,0 +1,50 @@
+/// Returns the lowercase equivalent of this string literal, as a new string literal.
+///
+/// See [`str::to_lowercase`](https://doc.rust-lang.org/std/primitive.str.html#method.to_lowercase).
+///
+/// # Examples
+///
+/// ```
+/// assert_eq!("hello", const_str::to_lowercase!("HELLO"));
+/// ```
+///
+#[macro_export]
+macro_rules! to_lowercase {
+    ($str:literal) => {
+        $crate::__imp::to_lowercase!($str)
+    };
+}
+
+/// Returns the uppercase equivalent of this string literal, as a new string literal.
+///
+/// See [`str::to_uppercase`](https://doc.rust-lang.org/std/primitive.str.html#method.to_uppercase).
+///
+/// # Examples
+///
+/// ```
+/// assert_eq!("HELLO", const_str::to_uppercase!("hello"));
+/// ```
+///
+#[macro_export]
+macro_rules! to_uppercase {
+    ($str:literal) => {
+        $crate::__imp::to_uppercase!($str)
+    };
+}
+
+/// Replaces all matches of a pattern with another string literal.
+///
+/// See [`str::replace`](https://doc.rust-lang.org/std/primitive.str.html#method.replace).
+///
+/// # Examples
+///
+/// ```
+/// assert_eq!("this is new", const_str::replace!("this is old", "old", "new"));
+/// ```
+///
+#[macro_export]
+macro_rules! replace {
+    ($str:literal, $from:literal, $to:literal) => {
+        $crate::__imp::replace!($str, $from, $to)
+    };
+}