+++ const-str/tests/test.rs
@@ -22,6 +22,8 @@ fn regex() {
     use regex::Regex;
     let re: &'static str = const_str::verified_regex!(r"^\d{4}-\d{2}-\d{2}$");
     assert!(Regex::new(re).is_ok());
+
+    const_str::regex_assert_match!(r"^\d{4}-\d{2}-\d{2}$", "2014-01-01");
 }
 
 #[cfg(feature = "verify-http")]