+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/Nugine/const-str/compare/v0.1.2...HEAD
 
+### Added
+
++ `regex_assert_match` (requires the feature `verify-regex`)
+
 ## [0.1.2] - 2020-09-11
 
 [0.1.2]: https://github.com/Nugine/const-str/tree/v0.1.2